build: Move memory/gdt.* to boot/
