+++ kernel/kernel.asm
@@ -20,7 +20,7 @@ section .multiboot
 
 %include "interrupt/interrupt.asm"
 %include "interrupt/handler.asm"
-%include "memory/gdt.asm"
+%include "include/asm/gdt.asm"
 
 
 extern kmain                ; it is convention to name main kernel
