+++ kernel/kernel.c
@@ -6,7 +6,7 @@
 #include "io/video/video.h"
 #include "io/serial/serial.h"
 #include "interrupt/interrupt.h"
-#include "memory/gdt.h"
+#include "boot/gdt.h"
 
 extern void kmain(multiboot_info_t *multiboot_info, uint32_t multiboot_magic)
 {
